requirements.txt
google-generativeai>=0.8.0
pdfminer.six>=20221105
pypdf>=3.17.0
pytest>=7.4.0
pytest-mock>=3.12.0
python-dotenv>=1.0.0